rebrand for Versity uses
We want to rename & rebrand the initial config to match our Versity uses.

The start changes are there to make it easier to run lots of these for prod,
testing, etc.

We also remove the upstream canned example charts and add the initial version
of our Versity Production monitoring charts. This is clearly a work in
progress, but shows where & how to get new dashboards added.